#3657: refactor page state
