+++ src/contentScript/messenger/api.ts
@@ -65,6 +65,7 @@ export const runEffectPipeline = getMethod("RUN_EFFECT_PIPELINE");
 export const runMapArgs = getMethod("RUN_MAP_ARGS");
 
 export const getPageState = getMethod("GET_PAGE_STATE");
+export const setPageState = getMethod("SET_PAGE_STATE");
 
 export const notify = {
   info: getNotifier("NOTIFY_INFO"),
